+++ packages/backfill-hasher/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - backfill-hasher
 
-<!-- This log was last generated on Fri, 12 Jun 2026 20:31:24 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 25 Jun 2026 20:21:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 6.7.11
+
+Thu, 25 Jun 2026 20:21:25 GMT
+
+### Patches
+
+- Bump workspace-tools to v0.41.10
+
+## 6.7.10
+
+Thu, 25 Jun 2026 02:50:11 GMT
+
+### Patches
+
+- Bump workspace-tools to v0.41.9
 
 ## 6.7.9