+++ packages/cache/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @lage-run/cache
 
-<!-- This log was last generated on Fri, 12 Jun 2026 20:31:24 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 25 Jun 2026 20:21:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.6.4
+
+Thu, 25 Jun 2026 20:21:25 GMT
+
+### Patches
+
+- Bump @lage-run/config to v0.9.14
+- Bump @lage-run/reporters to v1.8.3
+- Bump @lage-run/target-graph to v0.15.9
+
+## 1.6.3
+
+Thu, 25 Jun 2026 04:59:29 GMT
+
+### Patches
+
+- Bump @lage-run/config to v0.9.13
+
+## 1.6.2
+
+Thu, 25 Jun 2026 02:50:11 GMT
+
+### Patches
+
+- Bump @lage-run/config to v0.9.12
+- Bump @lage-run/reporters to v1.8.2
+- Bump @lage-run/target-graph to v0.15.8
 
 ## 1.6.1