+++ packages/config/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @lage-run/config
 
-<!-- This log was last generated on Fri, 12 Jun 2026 20:31:24 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 25 Jun 2026 20:21:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.9.14
+
+Thu, 25 Jun 2026 20:21:25 GMT
+
+### Patches
+
+- Bump @lage-run/runners to v1.4.12
+- Bump @lage-run/target-graph to v0.15.9
+- Bump workspace-tools to v0.41.10
+
+## 0.9.13
+
+Thu, 25 Jun 2026 04:59:29 GMT
+
+### Patches
+
+- Update dependency cosmiconfig to v9.0.2 (renovate@whitesourcesoftware.com)
+
+## 0.9.12
+
+Thu, 25 Jun 2026 02:50:11 GMT
+
+### Patches
+
+- Bump @lage-run/runners to v1.4.11
+- Bump @lage-run/target-graph to v0.15.8
+- Bump workspace-tools to v0.41.9
 
 ## 0.9.11